[UT] Fix unstable mv test cases (#40306)
